+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## **[Unrealeased]**
+- [Diagnostic + opam install](https://github.com/EruEri/kosu-lang/pull/80)
 - [Makefile install](https://github.com/EruEri/kosu-lang/pull/79)
 - [Road 0.5: base](https://github.com/EruEri/kosu-lang/pull/77):
     - Rewrite ast: